refactor(evaluation): precompute evaluator feedback keys at construction
Replace the __wrapped_evaluator_func__ breadcrumb (set on the normalized
wrapper, then unwrapped at extraction time) with an eager, best-effort parse
of the user's original function in DynamicRunEvaluator.__init__, stored as
self._feedback_keys. _extract_feedback_keys now reads that precomputed list
instead of re-parsing source, so only the execution wrapper is tracked at
runtime and key extraction no longer depends on functools.wraps propagating
a custom dunder through ensure_traceable.

Move _extract_code_evaluator_feedback_keys (and its AST helpers
_AST_STR_TYPES / _get_str_value) from _runner.py into evaluator.py, since it
is now consumed only there. This avoids an evaluator.py <-> _runner.py import
cycle and lets _runner.py drop its now-unused ast and textwrap imports.
