+++ python/langsmith/evaluation/_runner.py
@@ -2,7 +2,6 @@
 
 from __future__ import annotations
 
-import ast
 import collections
 import concurrent.futures as cf
 import functools
@@ -13,7 +12,6 @@ import logging
 import pathlib
 import queue
 import random
-import textwrap
 import threading
 import uuid
 from collections.abc import (
@@ -58,17 +56,6 @@ from langsmith.evaluation.evaluator import (
     comparison_evaluator,
     run_evaluator,
 )
-
-# Python 3.14+ removes ast.Str in favor of ast.Constant
-_AST_STR_TYPES: tuple = (
-    (ast.Str, ast.Constant) if hasattr(ast, "Str") else (ast.Constant,)
-)
-
-
-def _get_str_value(node: ast.expr) -> str:
-    """Get string value from ast.Str or ast.Constant."""
-    return node.value if isinstance(node, ast.Constant) else node.s  # type: ignore[return-value,union-attr,attr-defined]
-
 
 if TYPE_CHECKING:
     import pandas as pd
@@ -2236,132 +2223,14 @@ def _get_random_name() -> str:
 
 def _extract_feedback_keys(evaluator: RunEvaluator):
     if isinstance(evaluator, DynamicRunEvaluator):
-        if getattr(evaluator, "func", None):
-            return _extract_code_evaluator_feedback_keys(evaluator.func)
-        elif getattr(evaluator, "afunc", None):
-            return _extract_code_evaluator_feedback_keys(evaluator.afunc)
+        # Keys are precomputed from the user's original function at construction.
+        return getattr(evaluator, "_feedback_keys", None) or []
     # TODO: Support for DynamicComparisonRunEvaluator
     if hasattr(evaluator, "evaluator"):
         # LangChainStringEvaluator
         if getattr(getattr(evaluator, "evaluator"), "evaluation_name", None):
             return [evaluator.evaluator.evaluation_name]
     return []
-
-
-def _extract_code_evaluator_feedback_keys(func: Callable) -> list[str]:
-    # Unwrap to the user's original function so key extraction reads their source.
-    func = getattr(func, "__wrapped_evaluator_func__", func)
-    python_code = inspect.getsource(func)
-
-    def extract_dict_keys(node):
-        if isinstance(node, ast.Dict):
-            keys = []
-            key_value = None
-            for key, value in zip(node.keys, node.values):
-                if isinstance(key, _AST_STR_TYPES):
-                    key_str = _get_str_value(key)
-                    if key_str == "key" and isinstance(value, _AST_STR_TYPES):
-                        key_value = _get_str_value(value)
-            return [key_value] if key_value else keys
-        elif (
-            isinstance(node, ast.Call)
-            and isinstance(node.func, ast.Name)
-            and node.func.id == "dict"
-        ):
-            for keyword in node.keywords:
-                if keyword.arg == "key" and isinstance(keyword.value, _AST_STR_TYPES):
-                    return [_get_str_value(keyword.value)]
-        return []
-
-    def extract_evaluation_result_key(node):
-        if (
-            isinstance(node, ast.Call)
-            and isinstance(node.func, ast.Name)
-            and node.func.id == "EvaluationResult"
-        ):
-            for keyword in node.keywords:
-                if keyword.arg == "key" and isinstance(keyword.value, _AST_STR_TYPES):
-                    return [_get_str_value(keyword.value)]
-        return []
-
-    def extract_evaluation_results_keys(node, variables):
-        if (
-            isinstance(node, ast.Call)
-            and isinstance(node.func, ast.Name)
-            and node.func.id == "EvaluationResults"
-        ):
-            for keyword in node.keywords:
-                if keyword.arg == "results":
-                    if isinstance(keyword.value, ast.Name):
-                        return variables.get(keyword.value.id, [])
-                    elif isinstance(keyword.value, ast.List):
-                        keys = []
-                        for elt in keyword.value.elts:
-                            keys.extend(extract_evaluation_result_key(elt))
-                        return keys
-        elif isinstance(node, ast.Dict):
-            for key, value in zip(node.keys, node.values):
-                if isinstance(key, _AST_STR_TYPES) and _get_str_value(key) == "results":
-                    if isinstance(value, ast.List):
-                        keys = []
-                        for elt in value.elts:
-                            if isinstance(elt, ast.Dict):
-                                for elt_key, elt_value in zip(elt.keys, elt.values):
-                                    if (
-                                        isinstance(elt_key, _AST_STR_TYPES)
-                                        and _get_str_value(elt_key) == "key"
-                                    ):
-                                        if isinstance(elt_value, _AST_STR_TYPES):
-                                            keys.append(_get_str_value(elt_value))
-                            elif (
-                                isinstance(elt, ast.Call)
-                                and isinstance(elt.func, ast.Name)
-                                and elt.func.id in ("EvaluationResult", "dict")
-                            ):
-                                for keyword in elt.keywords:
-                                    if keyword.arg == "key" and isinstance(
-                                        keyword.value, _AST_STR_TYPES
-                                    ):
-                                        keys.append(_get_str_value(keyword.value))
-
-                        return keys
-        return []
-
-    python_code = textwrap.dedent(python_code)
-
-    try:
-        tree = ast.parse(python_code)
-        function_def = tree.body[0]
-        if not isinstance(function_def, (ast.FunctionDef, ast.AsyncFunctionDef)):
-            return []
-
-        variables = {}
-        keys = []
-
-        for node in ast.walk(function_def):
-            if isinstance(node, ast.Assign):
-                if isinstance(node.value, ast.List):
-                    list_keys = []
-                    for elt in node.value.elts:
-                        list_keys.extend(extract_evaluation_result_key(elt))
-                    if isinstance(node.targets[0], ast.Name):
-                        variables[node.targets[0].id] = list_keys
-            elif isinstance(node, ast.Return) and node.value is not None:
-                dict_keys = extract_dict_keys(node.value)
-                eval_result_key = extract_evaluation_result_key(node.value)
-                eval_results_keys = extract_evaluation_results_keys(
-                    node.value, variables
-                )
-
-                keys.extend(dict_keys)
-                keys.extend(eval_result_key)
-                keys.extend(eval_results_keys)
-
-        # If no keys found, return the function name
-        return keys if keys else [function_def.name]
-
-    except SyntaxError:
-        return []
 
 
 def _to_pandas(
